refactor:(stripe integration): minor cleanup of client dependencies
